requirements.txt
langchain
langchain_community
langchain_huggingface
langchain_groq
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers